+++ neu-admin-be/routes/partners.js
@@ -77,7 +77,7 @@ router.post(
     "/api/create-partner",
     initPartnerMiddleware,
     upload.array("docs"),
-    emptyPartnerFileInputValidation,
+
     async (req, res) => {
         try {
             const {
@@ -112,6 +112,34 @@ router.post(
             console.log(req.payload, "req.payload, create-partner api");
             const partnerId = req.payload;
             const docs = req.files;
+            if (!docs) {
+                console.log('ko có file được up lên create-partner api')
+            } else {
+                // trường hợp có file up lên, lưu các thông tin của các file up lên vào db
+                console.log(docs, "docs create-partner api")
+                const partnerDocsArray = docs.map((doc) => {
+                    return {
+                        docLink: doc.location,
+                        docName: doc.originalname,
+                        docSize: doc.size,
+                        docType: doc.mimetype,
+                        program: {
+                            id: programId,
+                        },
+                        partner: {
+                            id: partnerId,
+                        },
+                    };
+                });
+    
+                const savedPartnerDocsArray = await PartnerDocsSchema.insertMany(
+                    partnerDocsArray
+                );
+                console.log(
+                    savedPartnerDocsArray,
+                    "savedPartnerDocsArray create-partner api"
+                );
+            }
             console.log(docs, "req.files create-partner api");
             const newPartner = {
                 vn_name: vn_name,
@@ -151,28 +179,7 @@ router.post(
             );
             console.log(storingPartner, "storingPartner create-partner api");
 
-            const partnerDocsArray = docs.map((doc) => {
-                return {
-                    docLink: doc.location,
-                    docName: doc.originalname,
-                    docSize: doc.size,
-                    docType: doc.mimetype,
-                    program: {
-                        id: programId,
-                    },
-                    partner: {
-                        id: partnerId,
-                    },
-                };
-            });
 
-            const savedPartnerDocsArray = await PartnerDocsSchema.insertMany(
-                partnerDocsArray
-            );
-            console.log(
-                savedPartnerDocsArray,
-                "savedPartnerDocsArray create-partner api"
-            );
 
             res.json({ error: false, message: "Lưu thành công đối tác" });
         } catch (error) {
@@ -258,11 +265,12 @@ router.put(
             });
             const docsControlResult = docsControl(docs1Refs, partnerDocsArrayInDb);
             console.log(docsControlResult, "docsControlResult");
-            if (docsControlResult === "DOCS1_EMPTY_INPUT_ERROR") {
-                console.log("empty docs1 input");
-                docsControlResultError = "DOCS1_EMPTY_INPUT_ERROR";
-                // res.json({ error: true, message: "Hãy điền đầy đủ form" })
-            } else if (docsControlResult === "DUPLICATED_DOCS_ERROR") {
+            // if (docsControlResult === "DOCS1_EMPTY_INPUT_ERROR") {
+            //     console.log("empty docs1 input");
+            //     docsControlResultError = "DOCS1_EMPTY_INPUT_ERROR";
+            //     // res.json({ error: true, message: "Hãy điền đầy đủ form" })
+            // }
+            if (docsControlResult === "DUPLICATED_DOCS_ERROR") {
                 console.log("duplicated docs");
                 docsControlResultError = "DUPLICATED_DOCS_ERROR";
                 // res.json({ error: true, message: "có file trùng" })
@@ -338,8 +346,6 @@ router.put(
                     console.log("không có ảnh mới để lưu");
                 }
                 res.json({ error: false, message: "Đối tác đã được sửa thành công" });
-            } else if (docsControlResultError === "DOCS1_EMPTY_INPUT_ERROR") {
-                res.json({ error: true, message: "Hãy điền đầy đủ form" })
             } else if (docsControlResultError === "DUPLICATED_DOCS_ERROR") {
                 res.json({ error: true, message: "có file trùng" })
             }
@@ -355,17 +361,25 @@ router.delete("/api/delete-partner/:id", async (req, res) => {
     try {
         const { id } = req.params;
         console.log(id, "::id delete api::");
+        const checkPartnerDocs = await PartnerDocsSchema.find({ partner: { id: new ObjectId(id) } })
+        if (checkPartnerDocs.length === 0) {
+            console.log(`ko có file nào trong db của partner id: ${id}`)
+        } else {
+            // trường hợp có thông tin file trong db của partner đó
+            // Xóa các file cần xóa trên s3
+            const programId = checkPartnerDocs[0].program.id.toString()
+            // prefix là toàn bộ phần trước tên file
+            const prefix = `program-${programId}/partner-file-${id}/`;
+            const delFolderRes = await deleteFolder(prefix);
+            console.log(delFolderRes, "delFolderRes delete-partner api");
 
-        // Xóa các file cần xóa trên s3
-        const prefix = `partner-file-${id}`;
-        const delFolderRes = await deleteFolder(prefix);
-        console.log(delFolderRes, "delFolderRes delete-partner api");
-
-        // Xóa thông tin các file cần xóa trong db
-        const delPartnerFiles = await PartnerDocsSchema.deleteMany({
-            partner: { id: new ObjectId(id) },
-        });
-        console.log(delPartnerFiles, "delPartnerFiles delete-partner api");
+            // Xóa thông tin các file cần xóa trong db
+            const delPartnerFiles = await PartnerDocsSchema.deleteMany({
+                partner: { id: new ObjectId(id) },
+            });
+            console.log(delPartnerFiles, "delPartnerFiles delete-partner api");
+        }
+        
 
         // Xóa thông tin partner cần xóa
         const deletingPartner = await PartnerSchema.findOneAndDelete({ _id: id });
@@ -443,12 +457,12 @@ router.use((error, req, res, next) => {
     if (error) {
         console.log(error, "custom error handler");
 
-        if (
-            error.code === "EMPTY_PARTNER_FILE_INPUT_ERROR"
-        ) {
-            console.log(error.code, "empty input error");
-            return res.json({ error: true, message: "Chưa chọn file nào" });
-        }
+        // if (
+        //     error.code === "EMPTY_PARTNER_FILE_INPUT_ERROR"
+        // ) {
+        //     console.log(error.code, "empty input error");
+        //     return res.json({ error: true, message: "Chưa chọn file nào" });
+        // }
 
         // if (error.code === "PARTNER_INPUTS_TYPE_ERROR") {
         //     console.log("input type error");
